Move search-related stuff into Search module
